// apps/webapp/src/app/core/api/models/update-product-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface UpdateProductDto {

  /**
   * The name of the product
   */
  name?: string;

  /**
   * The price of the product
   */
  price?: number;

  /**
   * The current inventory quantity
   */
  quantity?: number;

  /**
   * The stock keeping unit identifier
   */
  sku?: string;
}